requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.2.0
countryinfo==0.1.2
cycler==0.12.1
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
filelock==3.12.4
fonttools==4.44.0
gevent==23.9.1
greenlet==3.0.1
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
platformdirs==3.11.0
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.5
websocket==0.2.1
websocket-client==1.7.0
zope.event==5.0
zope.interface==6.1
whitenoise==6.6.0
gunicorn==21.2.0
